+++ internal/core/ports/CartRepository.go
@@ -6,7 +6,7 @@ import (
 
 type CartRepository interface {
 	CreateCart(cart *domain.Cart) (*domain.Cart, error)
-	GetCart(id int) (*domain.Cart, error)
+	GetCart(id int64) (*domain.Cart, error)
 	UpdateCart(cart *domain.Cart) (*domain.Cart, error)
-	DeleteCart(id int) error
+	DeleteCart(id int64) error
 }
