+++ internal/core/ports/CategoryRepository.go
@@ -7,8 +7,8 @@ import (
 type CategoryRepository interface {
 	CreateCategory(category *domain.Category) (*domain.Category, error)
 	UpdateCategory(category *domain.Category) (*domain.Category, error)
-	DeleteCategory(id int) error
-	GetCategory(id int) (*domain.Category, error)
+	DeleteCategory(id int64) error
+	GetCategory(id int64) (*domain.Category, error)
 	GetCategoryByName(name string) *domain.Category
 	GetCategories() []domain.Category
 }
