+++ internal/core/ports/ProductRepository.go
@@ -6,8 +6,8 @@ import (
 
 type ProductRepository interface {
 	GetProducts() []domain.Product
-	GetProduct(id int) (*domain.Product, error)
-	DeleteProduct(id int) error
+	GetProduct(id int64) (*domain.Product, error)
+	DeleteProduct(id int64) error
 	CreateProduct(product *domain.Product) (*domain.Product, error)
 	UpdateProduct(product *domain.Product) (*domain.Product, error)
 }
