+++ internal/core/service/product_service.go
@@ -13,7 +13,7 @@ func NewProductService(repository ports.ProductRepository) *ProductService {
 	return &ProductService{repository: repository}
 }
 
-func (s *ProductService) GetProductById(id int) (*domain.Product, error) {
+func (s *ProductService) GetProductById(id int64) (*domain.Product, error) {
 	return s.repository.GetProduct(id)
 }
 
@@ -24,7 +24,7 @@ func (s *ProductService) UpdateProduct(product *domain.Product) (*domain.Product
 	return s.repository.UpdateProduct(product)
 }
 
-func (s *ProductService) DeleteProduct(id int) error {
+func (s *ProductService) DeleteProduct(id int64) error {
 	return s.repository.DeleteProduct(id)
 }
 
@@ -32,6 +32,6 @@ func (s *ProductService) GetAllProducts() []domain.Product {
 	return s.repository.GetProducts()
 }
 
-func (s *ProductService) GetProduct(id int) (*domain.Product, error) {
+func (s *ProductService) GetProduct(id int64) (*domain.Product, error) {
 	return s.repository.GetProduct(id)
 }
